in CuPy 13.3 there was implicit conversion from uint8->int16 on multiplication with scalar 1000.
For cupy 14.0 this dtype conversion does not happen and overflow occurs instead, so we need to declare as int16 to start with

+++ python/cucim/src/cucim/skimage/feature/tests/test_corner.py
@@ -264,8 +264,8 @@ def test_structure_tensor_eigenvalues(dtype):
 
 
 def test_structure_tensor_eigenvalues_3d():
-    cube9 = cp.ones((9,) * 3, dtype=cp.uint8)
-    cube7 = cp.ones((7,) * 3, dtype=cp.uint8)
+    cube9 = cp.ones((9,) * 3, dtype=cp.int16)
+    cube7 = cp.ones((7,) * 3, dtype=cp.int16)
     image = pad(cube9, 5, mode="constant") * 1000
     boundary = (
         pad(cube9, 5, mode="constant") - pad(cube7, 6, mode="constant")
